requirements.txt
fastapi
uvicorn
pdfplumber
langchain
sentence-transformers
faiss-cpu
pydantic
